backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
pydantic-settings
google-generativeai
python-multipart
groq
reportlab
textract
openai
langchain
